Set color before colorscheme.
This avoids a bug in vim <8 where the order changes the color
of the loaded colorscheme.  It seems to be fixed in vim8
which has resulted in different colors on the vim8 systems
vs older systems.  Fixing the order here makes everything
use the 'newer' vim8 colors for solarized which turn out
to match the advertised scheme.